update: separate styled components from Login and ChatPage
